Don't use category in query. Fix words backup.

+++ Makefile
@@ -1,5 +1,5 @@
 site/words.js: data/words.csv
-	mv $@ $@-$$(date --iso-8601=seconds)
+	if [ -f $@ ]; then mv $@ $@-$$(date --iso-8601=seconds); fi
 	API_KEY=$(API_KEY) node index.js > $@
 	echo ']' >> $@
